+++ apps/dev_project/src/index.tsx
@@ -19,8 +19,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import { AppDefaultAuthentication } from "./components/default_example/AppDefaultAuthentication.tsx";
-import AppAwsAuthentication from "./components/aws_test/AppAwsAuthentication.tsx";
-import AppUI from "./components/ui_test/AppUI.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>