+++ packages/shared/src/types/modules/ui/uiModuleInterfaces.ts
@@ -26,6 +26,12 @@ import {UICookieBannerProps} from "./cookieBanner/cookieBannerModuleInterfaces";
 import {NavbarOrchestratorProps} from "../../../modules/default_modules/ui/navbar/navbarOrchestrator";
 import {ContentBarOrchestratorProps} from "../../../modules/default_modules/ui/contentBar/contentBarOrchestrator";
 import {UIContentBarProps} from "./contentBar/contentBarModuleInterfaces";
+import {
+  CaseReducer,
+  PayloadAction,
+  Slice,
+  SliceCaseReducers,
+} from "@reduxjs/toolkit";
 
 export type UIState = {
   navbarCollapsed: boolean;
@@ -36,13 +42,10 @@ export type UIStoreState = {
   [MandatoryModuleNames.UI]: UIState;
 };
 
-export type UIExtras = {
-  setNavbarCollapsed: (payload: boolean) => {type: string; payload: boolean};
-  toggleNavbar: () => {type: string};
-  setCollapsible: (payload: boolean) => {type: string; payload: boolean};
-};
-
-export type UIModuleType<TState extends UIState = UIState> = {
+export type UIModuleType<
+  TState extends UIState = UIState,
+  TCaseReducers extends SliceCaseReducers<TState> = SliceCaseReducers<TState>,
+> = {
   UILayerHeader: React.ComponentType<
     UIHeaderProps & {uiComponent?: React.ComponentType<UIHeaderProps>}
   >;
@@ -57,5 +60,4 @@ export type UIModuleType<TState extends UIState = UIState> = {
   UILayerNavbar: React.ComponentType<
     NavbarOrchestratorProps & {uiComponent?: React.ComponentType<UINavbarProps>}
   >;
-  extras: UIExtras;
-} & FFStoreModule<TState>;
+} & FFStoreModule<TState, TCaseReducers>;