+++ packages/shared/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@iavofficial/frontend-framework-shared-v1.0.0-dev.14](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.13...@iavofficial/frontend-framework-shared-v1.0.0-dev.14) (2025-12-02)
+
+### Bug Fixes
+
+* Fixed Ui Module types ([47de40f](https://github.com/iavofficial/IAVFrontendFramework/commit/47de40f5996df2b057b25c5b419b8de8fda8356c))
+
 ## [@iavofficial/frontend-framework-shared-v1.0.0-dev.13](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.12...@iavofficial/frontend-framework-shared-v1.0.0-dev.13) (2025-12-01)
 
 ## [@iavofficial/frontend-framework-shared-v1.0.0-dev.12](https://github.com/iavofficial/IAVFrontendFramework/compare/@iavofficial/frontend-framework-shared-v1.0.0-dev.11...@iavofficial/frontend-framework-shared-v1.0.0-dev.12) (2025-11-28)